[TEST] Update commandParser testcases
 - testvector 디렉토리 내용 중복으로 삭제
 - input_20_20.txt 경로 수정
 - Fixture 내 validCommand / InvalidCommand 모두 생성하는 것으로 변경
