+++ deep_thinking/deep_thinking_test/deep_thinking_test.vcxproj
@@ -34,7 +34,7 @@
   <PropertyGroup Label="UserMacros" />
   <ItemGroup>
     <ClInclude Include="pch.h" />
-    <ClInclude Include="TC_CommandParser.h" />
+    <ClInclude Include="CommandParserTest.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="EmployeeInfoTest.cpp" />
@@ -47,10 +47,15 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="EmployeeManagerTest.cpp" />
-    <ClCompile Include="TC_CommandParser.cpp" />
+    <ClCompile Include="CommandParserTest.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\deep_thinking\deep_thinking.vcxproj">
+      <Project>{5b6a8931-80c3-41be-9727-cfa5dd5ecc3e}</Project>
+    </ProjectReference>
   </ItemGroup>
   <ItemDefinitionGroup />
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
